Debug panic on grid edges

+++ src/grid/directions.rs
@@ -1,7 +1,7 @@
 pub struct Directions {}
 
 impl Directions {
-    pub fn get_all_directions() -> Vec<(i32, i32)> {
+    pub fn get_all_directions() -> Vec<(isize, isize)> {
         vec![(-1, 0), (1, 0), (0, 1), (0, -1),(1, -1), (1, 1), (-1, -1), (-1, 1)]
     }
 }
